code/requirements.txt
google-genai
pandas
python-dotenv
